+++ bravado_types/render.py
@@ -13,14 +13,7 @@ DEFAULT_RESOURCE_TYPE_FORMAT = "{}Resource"
 DEFAULT_OPERATION_TYPE_FORMAT = "{}Operation"
 DEFAULT_MODEL_TYPE_FORMAT = "{}Model"
 
-
-class ResponseTypes(str, Enum):
-    success = 'success'
-    union = 'union'
-    any = 'any'
-
-
-DEFAULT_RESPONSE_TYPES = ResponseTypes.success
+DEFAULT_MODEL_INHERITANCE = False
 
 
 class ArrayTypes(str, Enum):
@@ -30,6 +23,15 @@ class ArrayTypes(str, Enum):
 
 
 DEFAULT_ARRAY_TYPES = ArrayTypes.list
+
+
+class ResponseTypes(str, Enum):
+    success = 'success'
+    all = 'all'
+    any = 'any'
+
+
+DEFAULT_RESPONSE_TYPES = ResponseTypes.success
 
 
 class RenderConfig:
@@ -44,7 +46,7 @@ class RenderConfig:
         model_type_format: str = None,
         array_types: ArrayTypes = None,
         response_types: ResponseTypes = None,
-        model_inheritance: bool = False,
+        model_inheritance: bool = None,
         custom_templates_dir: str = None,
         postprocessor: Callable[[str, str], Any] = None,
     ):
@@ -61,14 +63,13 @@ class RenderConfig:
             types in the schema.
             - list: Use typing.List. This is the default behavior.
             - sequence: Use typing.Sequence
-            - union: use a union of typing.List and typing.Tuple
-        :param response_types: ResponseTypes enum member indicating how
-            operation response types should be annotated. A value of 'success'
-            indicates that response types should be a union of defined response
-            types for 2xx status codes. A value of 'all' indicates that
-            response types should be a union of all documented response types.
-            A value of 'any' indicates that all operations should be annotated
-            as returning Any. If not specified, the default is 'success'.
+            - union: Use a union of typing.List and typing.Tuple
+        :param response_types: ResponseTypes member indicating how to represent
+            operation response types.
+            - success: Use a union of defined response types for 2xx status
+                       codes. This is the default behavior.
+            - all: Use a union of all defined response type.
+            - any: Use Any for all operation responses.
         :param model_inheritance: If True, the model type hierarchy will
             reflect model inheritance relationships as expressed by the allOf
             schema property. If False, model types will only inherit from
@@ -97,6 +98,8 @@ class RenderConfig:
         self.array_types = array_types or DEFAULT_ARRAY_TYPES
         self.response_types = response_types or DEFAULT_RESPONSE_TYPES
 
+        if model_inheritance is None:
+            model_inheritance = DEFAULT_MODEL_INHERITANCE
         self.model_inheritance = model_inheritance
 
         self.custom_templates_dir = custom_templates_dir
@@ -146,6 +149,7 @@ class RenderConfig:
 def render(metadata: Metadata, spec: SpecInfo, config: RenderConfig) -> None:
     """
     Render module and stub files for a given Swagger schema.
+    :param metadata: Code generation metadata.
     :param spec: SpecInfo representing the schema.
     :param config: Rendering config.
     """